Support sharing FSH content via Gist links (#70)
* Support sharing FSH content via Gist links

This adds support for sharing FSH content via Gists. This
allows for longer FSH to be shared, since a gist does not have
the same limit in size that is enforced on us by bitly. The down
side is that authors will have to make the gist.

* Refactor the dialogs to be separate

Instead of one single dialog, gists are now more of a secondary
option that is only accessible from the share dialog, or if we
fail to generate a shareable link directly.

* Add an extra newline between files

* Fix the failing tests

* Add a button to copy FSH to clipboard from gist modal

* Refactor copying in share modal

* Update modal size to be md not sm

* Make the gist modal medium sized as well for consistency