Support FHIR R5 (#134)
* Update to current us core version

* Update MUI theme to avoid console error

* R5 and automatic dependencies

- Support specifying and loading R5 package
- Support loading 'latest' versions of packages
- Add terminology package for any FHIR version
- Add extensions package when FHIR R5 is used

* Add automatic dependencies for GoFSH

* Refactor and marking dependencies as automatic

Small refactor to make dependencies an object with packageId and version.
This makes the code easier to read and to understand logically. It also
supports adding a flag to indicate which dependencies are "automatically
loaded" so that it can be included in the error message if a "latest" version
cannot be identified and downloaded.

A few other places that relied on for loops were refactored to use .forEach
and .map in order to follow best practices. A few variable names were changed
to be more clear and simple.

* Add override to avoid error when dev server restarts (during development only)

* Don't add automatic dependency if any version was already specified